feat: make changes based on feedback
